requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
shap>=0.45.0
streamlit>=1.37.0
opencv-python>=4.9.0.80
requests>=2.31.0
